move simulator to a pop-over
